refactor using classes to get urls to the api
This adds two new classes that know which urls to use to get to
specific endpoints of the DHL EFulfillment sandox and production APIs.

This also adds a config value that makes possible to select one of
the two environments easily.

This also refactors the spec for dhl_efulfillment.rb.